services/twin-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
